Initial version of HSP.HTML
This module aims to provide well-form HTML 4.01 output. It currently
does not attempt to validate the document in anyway. So, it will not
detect invalid tags and attributes or missing required fields.

The rendering is similar to HSP.XML but with the following changes:

 * no short tags are used, e.g., <script></script> instead of <script />
 * the end tag is forbidden for some elements, for these we:
   * render only the open tag, e.g., <br>
   * throw an error if the tag contains children
 * optional end tags are always rendered

This thread discusses the reasons behind the creation of this module.

http://groups.google.com/group/haskell-server-pages/browse_thread/thread/b27c1d8e7326eb1c

I expect additional updates will follow.