renderAsHTML: render <script> and <style> content as #CDATA. Don't raise errors on invalid children.
In XML, the contents of <script> and <style> are #PCDATA, but in HTML they are #CDATA. So, we unescape the contents of those tags before outputing them in HTML mode.

Additionally, instead of raising errors when unexpected children are found (e.g., <img> having a child), we output the children anyway. Since renderAsHTML does not do any other 
validation, this seems in line with the way the rest of the function works. If you want validation, it is better to run it through a separate HTML validator.

See these threads for more information:

http://groups.google.com/group/haskell-server-pages/browse_thread/thread/70216bd746ae1ec4
http://groups.google.com/group/haskell-server-pages/browse_thread/thread/171a1e2372b97020